pyproject.toml
[project]
name = "ai-logs-review"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "faiss-cpu (>=1.11.0.post1,<2.0.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "gpt4all (>=2.8.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
